src/modules/rentals/useCases/createRental/createRentalUseCase.ts
import { inject, injectable } from "tsyringe";

import { ICarsRepository } from "@modules/cars/repositories/ICarsRepository";
import { Rental } from "@modules/rentals/infra/typeorm/entities/rental";
import { IRentalsRepository } from "@modules/rentals/repositories/IRentalsRepository";
import { IDateProvider } from "@shared/container/providers/dateProvider/IDateProvider";
import { AppError } from "@shared/errors/appError";

interface IRequest {
  car_id: string;
  user_id: string;
  expected_return_date: Date;
}

@injectable()
class CreateRentalUseCase {
  constructor(
    @inject("RentalsRepository")
    private rentalsRepository: IRentalsRepository,

    @inject("DayjsDateProvider")
    private dateProvider: IDateProvider,

    @inject("CarsRepository")
    private carsRepository: ICarsRepository,
  ) {}

  async execute({
    car_id,
    expected_return_date,
    user_id,
  }: IRequest): Promise<Rental> {
    const minimumHour = 24;

    const unavailableCar = await this.rentalsRepository.findRentalOpenByCar(
      car_id,
    );

    if (unavailableCar) {
      throw new AppError("car is unavailable");
    }

    const rentalOpenToUser = await this.rentalsRepository.findRentalOpenByUser(
      user_id,
    );

    if (rentalOpenToUser) {
      throw new AppError("There's a rental in progress for user");
    }

    const dateNowFormat = this.dateProvider.dateNow();

    const compare = this.dateProvider.compareInHours(
      dateNowFormat,
      expected_return_date,
    );

    if (compare < minimumHour) {
      throw new AppError("Invalid return time");
    }

    await this.carsRepository.updateAvailable(car_id, false);

    const rental = await this.rentalsRepository.create({
      car_id,
      user_id,
      expected_return_date,
    });

    return rental;
  }
}

export { CreateRentalUseCase };
